Table support

+++ src/doc-struct/location.h
@@ -1,13 +1,14 @@
 #pragma once
 
 #include "data/str.h"
+#include <stddef.h>
 
 typedef struct
 {
-	int first_line;
-	int first_column;
-	int last_line;
-	int last_column;
+	size_t first_line;
+	size_t first_column;
+	size_t last_line;
+	size_t last_column;
 	Str* src_file;
 } Location;
